Make diff more modular
