+++ caldir-cli/src/diff/calendar_diff.rs
@@ -1,129 +1,9 @@
 use anyhow::Result;
-use caldir_core::Event;
 use owo_colors::OwoColorize;
 use std::collections::HashMap;
-use std::fmt;
 
 use crate::calendar::Calendar;
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum DiffKind {
-    Create,
-    Update,
-    Delete,
-}
-
-impl fmt::Display for DiffKind {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            DiffKind::Create => write!(f, "+"),
-            DiffKind::Update => write!(f, "~"),
-            DiffKind::Delete => write!(f, "-"),
-        }
-    }
-}
-
-impl DiffKind {
-    /// Colorize text according to this diff kind
-    pub fn colorize(&self, text: &str) -> String {
-        match self {
-            DiffKind::Create => text.green().to_string(),
-            DiffKind::Update => text.yellow().to_string(),
-            DiffKind::Delete => text.red().to_string(),
-        }
-    }
-
-    /// Render the symbol with appropriate color
-    pub fn render(&self) -> String {
-        self.colorize(&self.to_string())
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct EventDiff {
-    pub kind: DiffKind,
-    pub old: Option<Event>,
-    pub new: Option<Event>,
-}
-
-impl fmt::Display for EventDiff {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}: {}", self.kind, self.event())
-    }
-}
-
-impl EventDiff {
-    pub fn get_diff(old_event: Option<Event>, new_event: Option<Event>) -> Option<EventDiff> {
-        match (&old_event, &new_event) {
-            (None, Some(_)) => Some(EventDiff {
-                kind: DiffKind::Create,
-                old: None,
-                new: new_event,
-            }),
-            (Some(_), None) => Some(EventDiff {
-                kind: DiffKind::Delete,
-                old: old_event,
-                new: None,
-            }),
-            (Some(old), Some(new)) => {
-                if old == new {
-                    None
-                } else {
-                    Some(EventDiff {
-                        kind: DiffKind::Update,
-                        old: old_event,
-                        new: new_event,
-                    })
-                }
-            }
-            (None, None) => None,
-        }
-    }
-
-    /// Get the event (prefer new, fallback to old)
-    fn event(&self) -> &Event {
-        self.new
-            .as_ref()
-            .or(self.old.as_ref())
-            .expect("EventDiff must have at least one event")
-    }
-
-    pub fn render(&self) -> String {
-        let event = self.event();
-        let summary = self.kind.colorize(&event.to_string());
-        let time = event.render_event_time();
-
-        format!("{} {} {}", self.kind.render(), summary, time.dimmed())
-    }
-}
-
-pub struct DiffBatch<'a>(pub Vec<CalendarDiff<'a>>);
-
-impl DiffBatch<'_> {
-    fn count_by_kind<'a>(diffs: impl Iterator<Item = &'a EventDiff>) -> (usize, usize, usize) {
-        let mut created = 0;
-        let mut updated = 0;
-        let mut deleted = 0;
-
-        for diff in diffs {
-            match diff.kind {
-                DiffKind::Create => created += 1,
-                DiffKind::Update => updated += 1,
-                DiffKind::Delete => deleted += 1,
-            }
-        }
-
-        (created, updated, deleted)
-    }
-
-    pub fn pull_counts(&self) -> (usize, usize, usize) {
-        Self::count_by_kind(self.0.iter().flat_map(|d| &d.to_pull))
-    }
-
-    pub fn push_counts(&self) -> (usize, usize, usize) {
-        Self::count_by_kind(self.0.iter().flat_map(|d| &d.to_push))
-    }
-}
+use crate::diff::{DiffKind, EventDiff};
 
 pub struct CalendarDiff<'a> {
     pub calendar: &'a Calendar,
