+++ caldir-cli/src/diff/mod.rs
@@ -1,0 +1,10 @@
+mod batch_diff;
+mod calendar_diff;
+mod diff_kind;
+mod event_diff;
+
+// Re-export:
+pub use batch_diff::BatchDiff;
+pub use calendar_diff::CalendarDiff;
+pub use diff_kind::DiffKind;
+pub use event_diff::EventDiff;
